### 2018-08-05 #873 @wholmgren
Re reviewers... perhaps the reviewers/authors in #825 (Calliope) would be appropriate.

cc @cwhanse @mikofski
