### 2018-08-05 #873 @katyhuff
Happy to edit this one -- it looks very interesting! Thanks for your submission @wholmgren . I agree with your assessment of appropriate reviewers.

@ecotillasanchez  @mdoucet  @sjpfenninger Would you be interested in reviewing [this JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.00873/joss.00873/10.21105.joss.00873.pdf)?
