### 2018-08-05 a325925
Document fix in release notes
