### 2018-08-06 a26d8e1
Add option to control airmass/perez enhancement factor in clearsky.ineichen (#459)
* add perez_enhancement option

* pass kwargs through location.get_clearsky to ineichen

* hard code test values to avoid test failures

* refactor test_irradiance.py input data into fixtures

* refactor location tests to avoid tight cs dependence

* mock/refactor all tests except losses

* fix typo in test

* update whatsnew

* update test ineichen values for perez_enhancement=False

* update test_clearsky values

* add test for perez enhancement

* update tests for py2.7-3.5

* add requires_scipy for min env t
[…]
