+++ pvlib/test/test_modelchain.py
@@ -1,4 +1,9 @@
 import sys
+try:
+    from unittest.mock import ANY
+except ImportError:
+    # python 2
+    from mock import ANY
 
 import numpy as np
 import pandas as pd
@@ -119,7 +124,7 @@ def test_run_model(system, location):
 
     expected = pd.Series(np.array([  183.522449305,  -2.00000000e-02]),
                          index=times)
-    assert_series_equal(ac, expected, check_less_precise=2)
+    assert_series_equal(ac, expected, check_less_precise=1)
 
 
 def test_run_model_with_irradiance(system, location):
@@ -159,132 +164,140 @@ def test_run_model_gueymard_perez(system, location):
     assert_series_equal(ac, expected)
 
 
-@requires_scipy
-def test_run_model_with_weather(system, location):
+def test_run_model_with_weather(system, location, weather, mocker):
     mc = ModelChain(system, location)
-    times = pd.date_range('20160101 1200-0700', periods=2, freq='6H')
-    weather = pd.DataFrame({'wind_speed':5, 'temp_air':10}, index=times)
-    ac = mc.run_model(times, weather=weather).ac
+    m = mocker.spy(system, 'sapm_celltemp')
+    weather['wind_speed'] = 5
+    weather['temp_air'] = 10
+    mc.run_model(weather.index, weather=weather)
+    assert m.call_count == 1
+    # assert_called_once_with cannot be used with series, so need to use
+    # assert_series_equal on call_args
+    assert_series_equal(m.call_args[0][1], weather['wind_speed'])  # wind
+    assert_series_equal(m.call_args[0][2], weather['temp_air'])  # temp
+    assert not mc.ac.empty
 
-    expected = pd.Series(np.array([  201.691634921,  -2.00000000e-02]),
-                         index=times)
-    assert_series_equal(ac, expected, check_less_precise=2)
 
-
-@requires_scipy
-def test_run_model_tracker(system, location):
+def test_run_model_tracker(system, location, weather, mocker):
     system = SingleAxisTracker(module_parameters=system.module_parameters,
                                inverter_parameters=system.inverter_parameters)
+    mocker.spy(system, 'singleaxis')
     mc = ModelChain(system, location)
-    times = pd.date_range('20160101 1200-0700', periods=2, freq='6H')
-    ac = mc.run_model(times).ac
-
-    expected = pd.Series(np.array([119.067713606,  nan]),
-                         index=times)
-    assert_series_equal(ac, expected, check_less_precise=2)
-
-    expect = pd.DataFrame(np.
-        array([[ 54.82513187,  90.        ,  11.0039221 ,  11.0039221 ],
-               [         nan,   0.        ,   0.        ,          nan]]),
-        columns=['aoi', 'surface_azimuth', 'surface_tilt', 'tracker_theta'],
-        index=times)
-    expect = expect[['tracker_theta', 'aoi', 'surface_azimuth', 'surface_tilt']]
-    assert_frame_equal(mc.tracking, expect, check_less_precise=2)
+    mc.run_model(weather.index, weather=weather)
+    assert system.singleaxis.call_count == 1
+    assert (mc.tracking.columns == ['tracker_theta', 'aoi', 'surface_azimuth',
+                                    'surface_tilt']).all()
+    assert mc.ac[0] > 0
+    assert np.isnan(mc.ac[1])
 
 
 def poadc(mc):
     mc.dc = mc.total_irrad['poa_global'] * 0.2
[…]
