+++ pvlib/test/test_irradiance.py
@@ -9,29 +9,58 @@ from numpy.testing import assert_almost_equal, assert_allclose
 
 from pandas.util.testing import assert_frame_equal, assert_series_equal
 
-from pvlib.location import Location
-from pvlib import clearsky
-from pvlib import solarposition
 from pvlib import irradiance
-from pvlib import atmosphere
 
-from conftest import (requires_ephem, requires_numba, needs_numpy_1_10,
-                      pandas_0_22)
+from conftest import (needs_numpy_1_10, pandas_0_22, requires_ephem,
+                      requires_numba)
 
-# setup times and location to be tested.
-tus = Location(32.2, -111, 'US/Arizona', 700)
 
-# must include night values
-times = pd.date_range(start='20140624', freq='6H', periods=4, tz=tus.tz)
+# fixtures create realistic test input data
+# test input data generated at Location(32.2, -111, 'US/Arizona', 700)
+# test input data is hard coded to avoid dependencies on other parts of pvlib
 
-ephem_data = solarposition.get_solarposition(
-    times, tus.latitude, tus.longitude, method='nrel_numpy')
 
-irrad_data = tus.get_clearsky(times, model='ineichen', linke_turbidity=3)
+@pytest.fixture
+def times():
+    # must include night values
+    return pd.date_range(start='20140624', freq='6H', periods=4,
+                         tz='US/Arizona')
 
-dni_et = irradiance.extraradiation(times.dayofyear)
 
-ghi = irrad_data['ghi']
+@pytest.fixture
+def irrad_data(times):
+    return pd.DataFrame(np.array(
+        [[   0.        ,    0.        ,    0.        ],
+         [  79.73860422,  316.1949056 ,   40.46149818],
+         [1042.48031487,  939.95469881,  118.45831879],
+         [ 257.20751138,  646.22886049,   62.03376265]]),
+        columns=['ghi', 'dni', 'dhi'], index=times)
+
+
+@pytest.fixture
+def ephem_data(times):
+    return pd.DataFrame(np.array(
+        [[124.0390863 , 124.0390863 , -34.0390863 , -34.0390863 ,
+          352.69550699,  -2.36677158],
+         [ 82.85457044,  82.97705621,   7.14542956,   7.02294379,
+           66.71410338,  -2.42072165],
+         [ 10.56413562,  10.56725766,  79.43586438,  79.43274234,
+          144.76567754,  -2.47457321],
+         [ 72.41687122,  72.46903556,  17.58312878,  17.53096444,
+          287.04104128,  -2.52831909]]),
+        columns=['apparent_zenith', 'zenith', 'apparent_elevation',
+                 'elevation', 'azimuth', 'equation_of_time'],
+        index=times)
+
+
+@pytest.fixture
+def dni_et(times):
+    return irradiance.extraradiation(times.dayofyear)
+
+
+@pytest.fixture
+def relative_airmass(times):
+    return pd.Series([np.nan, 7.58831596, 1.01688136, 3.27930443], times)
 
 
 # setup for et rad test. put it here for readability
@@ -63,9 +92,11 @@ def test_extraradiation(input, expected, method):
 
 
 @requires_numba
-def test_extraradiation_nrel_numba():
-    result = irradiance.extraradiation(times, method='nrel', how='numba', numthreads=8)
-    assert_allclose(result, [1322.332316, 1322.296282, 1322.261205, 1322.227091])
+def test_extraradiation_nrel_numba(times):
+    result = irradiance.extraradiation(times, method='nrel', how='numba',
+                                       numthreads=4)
+    assert_allclose(result,
[…]
