### 2018-08-06 09126e9
add python 3.7 to test matrix, documentation, setup.py (#520)
* add python 3.7 to travis ci

* minor updates to other configs

* update travis config

* update travis config

* update travis config

* update documentation

* update setup.py for 3.7
