### 2018-08-06 0326a6a
update get_total_irradiance arg names
