### 2018-08-06 1353611
add test values, fix botched merges
