+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -32,8 +32,7 @@ API Changes
 * Deprecated irradiance.total_irrad. Use irradiance.get_total_poa_irradiance
   instead. irradiance.total_irrad will be removed in 0.7. (:issue:`427`)
 * Removed 'klutcher' from get_sky_diffuse/total_irrad. This misspelling was
-  deprecated long ago but never removed.
-* pvsystem.calcparams_desoto now requires arguments for each module model parameter.
+  deprecated long ago but never removed. (:issue:`97`)
 * pvsystem.calcparams_desoto now requires arguments for each module model
   parameter. (:issue:`462`)
 * Add losses_parameters attribute to PVSystem objects and remove the kwargs
