### 2018-08-06 7479bc8
make test_forecast.py more robust, print problems instead of fail
