+++ pvlib/test/test_forecast.py
@@ -2,6 +2,7 @@ from datetime import datetime, timedelta
 import inspect
 from math import isnan
 from pytz import timezone
+import warnings
 
 import numpy as np
 import pandas as pd
@@ -53,11 +54,10 @@ def model(request):
     try:
         raw_data = amodel.get_data(_latitude, _longitude, _start, _end)
     except Exception as e:
-        print('Exception getting data for ', amodel)
-        print('latitude, longitude, start, end = ',
-              _latitude, _longitude, _start, _end)
-        print(e)
-        raw_data = pd.DataFrame(tz=_start.tzinfo)
+        warnings.warn('Exception getting data for {}.\n'
+                      'latitude, longitude, start, end = {} {} {} {}\n{}'
+                      .format(amodel, _latitude, _longitude, _start, _end, e))
+        raw_data = pd.DataFrame()  # raw_data.empty will be used later
     amodel.raw_data = raw_data
     return amodel
 
@@ -66,15 +66,16 @@ def model(request):
 def test_process_data(model):
     for how in ['liujordan', 'clearsky_scaling']:
         if model.raw_data.empty:
-            print('Could not process {} data with how={} '
-                  'because raw_data was empty'.format(model, how))
+            warnings.warn('Could not test {} process_data with how={} '
+                          'because raw_data was empty'.format(model, how))
             continue
         data = model.process_data(model.raw_data, how=how)
         for variable in _nonnan_variables:
             try:
                 assert not data[variable].isnull().values.any()
             except AssertionError:
-                print(model, variable, 'data contained null values')
+                warnings.warn('{}, {}, data contained null values'
+                              .format(model, variable))
 
 
 @requires_siphon
