### 2018-08-07 #873 @katyhuff
@ecotillasanchez @mdoucet Would either of you be interested in reviewing [this JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.00873/joss.00873/10.21105.joss.00873.pdf) ? It seems to be well within your area of expertise.  "pvlib python: a python package for modeling solar energy systems"
