+++ pvlib/test/conftest.py
@@ -1,4 +1,3 @@
-import sys
 import platform
 
 from pkg_resources import parse_version
@@ -7,7 +6,7 @@ import numpy as np
 import pytest
 
 
-skip_windows = pytest.mark.skipif('win' in sys.platform,
+skip_windows = pytest.mark.skipif(platform.system() == 'Windows',
                                   reason='does not run on windows')
 
 try:
