### 2018-08-07 #884 @whedon
[REVIEW]: pvlib python: a python package for modeling solar energy systems
**Submitting author:** @wholmgren (<a href="http://orcid.org/0000-0001-6218-9767">William F. Holmgren</a>)
**Repository:** <a href="https://github.com/pvlib/pvlib-python" target ="_blank">https://github.com/pvlib/pvlib-python</a>
**Version:** v0.5.2
