### 2018-08-08 2a436c6
ENH: implementing pvsyst recombination loss current for CdTe and a:Si (#504)
* implementing pvsyst recombination loss current for CdTe and a:Si

* closes #163
* add constant VOLTAGE_BUILTIN = 0.9 (V) to singlediode_methods.py
* add arguments d2mutau=0, voltage_builtin=VOLTAGE_BUILTIN, and
cells_in_series to calculate i_recomb and v_recomb
* add check for d2mutau to calc i_recomb, v_recomb, and gradients
* add terms to current and gradients

* ENH: need to set grad_i_recomb if d2mutau not given

* also set v_recomb to Inf, so that 1/v_recomb is always zero
* but precalculate grad_2i_recomb only
[…]
