### 2018-08-08 556badf
update to latest mpl code with minor mods
