+++ pvlib/_deprecation.py
@@ -101,7 +101,17 @@ Licensee agrees to be bound by the terms and conditions of this License
 Agreement.
 """
 
+# modified from Matplotlib b97cd2d (post 2.2.2) in the following ways:
+# 1. use module-level _projectName = 'pvlib' and
+#    _projectWarning = 'pvlibDeprecationWarning' in place of MPL specific
+#    string/Class.
+# 2. remove keyword only argument requirement for removal
+# 3. remove deprecated obj_type from deprecated function
+# 4. if removal is empty, say 'soon' instead of assuming two minor releases
+#    later.
+
 import functools
+import textwrap
 import warnings
 
 
@@ -119,46 +129,50 @@ class pvlibDeprecationWarning(UserWarning):
     pass
 
 
-pvlibDeprecation = pvlibDeprecationWarning
+# make it easier for others to copy paste this code into their projects
+_projectName = 'pvlib'
+_projectWarning = pvlibDeprecationWarning
 
 
-def _generate_deprecation_message(since, message='', name='',
-                                  alternative='', pending=False,
-                                  obj_type='attribute',
-                                  addendum=''):
+def _generate_deprecation_message(
+        since, message='', name='', alternative='', pending=False,
+        obj_type='attribute', addendum='', removal=''):
+
+    if removal == "":
+        removal = "soon"
+    elif removal:
+        if pending:
+            raise ValueError(
+                "A pending deprecation cannot have a scheduled removal")
+        removal = "in {}".format(removal)
 
     if not message:
+        message = (
+            "The %(name)s %(obj_type)s"
+            + (" will be deprecated in a future version"
+               if pending else
+               (" was deprecated in %(projectName)s %(since)s"
+                + (" and will be removed %(removal)s"
+                   if removal else
+                   "")))
+            + "."
+            + (" Use %(alternative)s instead." if alternative else "")
+            + (" %(addendum)s" if addendum else ""))
 
-        if pending:
-            message = (
-                'The {} {} will be deprecated in a '
-                'future version.'.format(name, obj_type))
-        else:
-            message = (
-                'The {} {} was deprecated in version '
-                '{}.'.format(name, obj_type, since))
-
-    altmessage = ''
-    if alternative:
-        altmessage = ' Use {} instead.'.format(alternative)
-
-    message = message + altmessage
-
-    if addendum:
-        message += addendum
-
-    return message
+    return message % dict(
+        func=name, name=name, obj_type=obj_type, since=since, removal=removal,
+        alternative=alternative, addendum=addendum, projectName=_projectName)
 
 
-def warn_deprecated(since, message='', name='', alternative='', pending=False,
-                    obj_type='attribute', addendum=''):
-    """Display deprecation warning in a standard way.
-
+def warn_deprecated(
+        since, message='', name='', alternative='', pending=False,
+        obj_type='attribute', addendum='', *, removal=''):
+    """
+    Used to display deprecation in a standard way.
     Parameters
     ----------
     since : str
         The release at which this API became deprecated.
-
     message : str, optional
         Override the default deprecation message.  The format
         specifier `%(name)s` may be used for the name of the function,
@@ -166,121 +180,109 @@ def warn_deprecated(since, message='', name='', alternative='', pending=False,
         to insert the name of an alternative to the deprecated
         function.  `%(obj_type)s` may be used to insert a friendly name
         for the type of object being deprecated.
-
     name : str, optional
         The name of the deprecated object.
-
     alternative : str, optional
-        An alternative function that the user may use in place of the
-        deprecated function.  The deprecation warning will tell the user
-        about this alternative if provided.
-
+        An alternative API that the user may use in place of the deprecated
+        API.  The deprecation warning will tell the user about this alternative
+        if provided.
     pending : bool, optional
         If True, uses a PendingDeprecationWarning instead of a
[…]
