+++ pvlib/_deprecation.py
@@ -166,7 +166,7 @@ def _generate_deprecation_message(
 
 def warn_deprecated(
         since, message='', name='', alternative='', pending=False,
-        obj_type='attribute', addendum='', *, removal=''):
+        obj_type='attribute', addendum='', removal=''):
     """
     Used to display deprecation in a standard way.
     Parameters
