### 2018-08-08 fe079ea
absolute_airmass becomes get_absolute_airmass for consistency and future proofing
