+++ docs/sphinx/source/api.rst
@@ -102,7 +102,7 @@ Airmass and atmospheric models
    :toctree: generated/
 
    location.Location.get_airmass
-   atmosphere.absolute_airmass
+   atmosphere.get_absolute_airmass
    atmosphere.get_relative_airmass
    atmosphere.pres2alt
    atmosphere.alt2pres
