### 2018-08-08 611f409
MAINT: rename singlediode_methods.py to just singlediode.py
* closes #515
* import pvlib in pvsystem and use pvlib.singlediode to differentiate
from existing pvsystem.singldiode() method already there!
* rename test_singlediode_methods.py to test_singlediode.py
* update documentation and comments
