### 2018-08-08 1a42153
DOC: move proof that estimated Voc is in q4
* closes #518
* add singlediode.rst and add to index
