+++ docs/sphinx/source/index.rst
@@ -81,6 +81,7 @@ Contents
    api
    comparison_pvlib_matlab
    variables_style_rules
+   singlediode
 
 
 Indices and tables
