### 2018-08-08 a49e959
DOC: add DOI to Jain and Kapoor
* proofed the docs and they are ok
