+++ docs/sphinx/source/singlediode.rst
@@ -14,10 +14,10 @@ a ``method`` keyword to the :func:`pvlib.pvsystem.singlediode` function:
 
 Lambert W-Function
 ------------------
-When ``method='lambertw'``, the Lambert W-Function is used as previously shown
-by Jain and Kapoor [1]. The following algorithm can be found on
+When ``method='lambertw'``, the Lambert W-function is used as previously shown
+by Jain and Kapoor [1, 2]. The following algorithm can be found on
 `Wikipedia: Theory of Solar Cells
-<https://en.wikipedia.org/wiki/Theory_of_solar_cells>`_ given the basic single
+<https://en.wikipedia.org/wiki/Theory_of_solar_cells>`_, given the basic single
 diode model equation.
 
 .. math::
@@ -47,7 +47,7 @@ The the module current can be solved using the Lambert W-function.
 
 Bishop's Algorithm
 ------------------
-The function :func:`pvlib.singlediode.bishop88` uses an explicit solution [2]
+The function :func:`pvlib.singlediode.bishop88` uses an explicit solution [3]
 that finds points on the IV curve by first solving for pairs :math:`(V_d, I)`
 where :math:`V_d` is the diode voltage :math:`V_d = V + I*Rs`. Then the voltage
 is backed out from :math:`V_d`. Points with specific voltage, such as open
@@ -89,10 +89,16 @@ desired condition.
 
 References
 ----------
-[1] A. Jain, A. Kapoor, "Exact analytical solutions of the
-parameters of real solar cells using Lambert W-function", Solar
-Energy Materials and Solar Cells, 81 (2004) 269-277.
+[1] "Exact analytical solutions of the parameters of real solar cells using
+Lambert W-function," A. Jain, A. Kapoor, Solar Energy Materials and Solar Cells,
+81, (2004) pp 269-277.
+:doi:`10.1016/j.solmat.2003.11.018`
 
-[2] "Computer simulation of the effects of electrical mismatches in
+[2] "A new method to determine the diode ideality factor of real solar cell
+using Lambert W-function," A. Jain, A. Kapoor, Solar Energy Materials and Solar
+Cells, 85, (2005) 391-396.
+:doi:`10.1016/j.solmat.2004.05.022`
+
+[3] "Computer simulation of the effects of electrical mismatches in
 photovoltaic cell interconnection circuits" JW Bishop, Solar Cell (1988)
-https://doi.org/10.1016/0379-6787(88)90059-2
+:doi:`10.1016/0379-6787(88)90059-2`
