### 2018-08-09 4c595ac
update extraradiation call in forecast.py
