### 2018-08-09 9fef243
update docs to use new api
