+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -84,6 +84,8 @@ Documentation
 * Updated several incorrect statements in ModelChain documentation regarding
   implementation status and default values. (:issue:`480`)
 * Expanded general contributing and pull request guidelines.
+* Added section on single diode equation with some detail on solutions used in
+  pvlib-python (:issue:`518`)
 
 
 Testing
