+++ docs/sphinx/source/singlediode.rst
@@ -1,21 +1,23 @@
 .. _singlediode:
 
-Single Diode Model
-==================
+Single Diode Equation
+=====================
 
-This section reviews the solutions to the single diode model used in
+This section reviews the solutions to the single diode equation used in
 pvlib-python to generate an IV curve of a PV module.
 
-pvlib-python supports two ways to solve the single diode model, by passing the
-a ``method`` keyword to the :func:`pvlib.pvsystem.singlediode` function:
+pvlib-python supports two ways to solve the single diode equation:
 
 1. Lambert W-Function
 2. Bishop's Algorithm
 
+The :func:`pvlib.pvsystem.singlediode` function allows the user to choose the
+method using the ``method`` keyword.
+
 Lambert W-Function
 ------------------
 When ``method='lambertw'``, the Lambert W-function is used as previously shown
-by Jain and Kapoor [1, 2]. The following algorithm can be found on
+by Jain and Kapoor [1-3]. The following algorithm can be found on
 `Wikipedia: Theory of Solar Cells
 <https://en.wikipedia.org/wiki/Theory_of_solar_cells>`_, given the basic single
 diode model equation.
@@ -28,8 +30,9 @@ diode model equation.
 Lambert W-function is the inverse of the function
 :math:`f \left( w \right) = w \exp \left( w \right)` or
 :math:`w = f^{-1} \left( w \exp \left( w \right) \right)` also given as
-:math:`w = W \left( w \exp \left( w \right) \right)`. Rearranging the single
-diode equation above with a Lambert W-function yields the following.
+:math:`w = W \left( w \exp \left( w \right) \right)`. Defining the following
+parameter, :math:`z`, is necessary to transform the single diode equation into
+a form that can be expressed as a Lambert W-function.
 
 .. math::
 
@@ -37,17 +40,18 @@ diode equation above with a Lambert W-function yields the following.
        \frac{R_s \left( I_L + I_0 \right) + V}{n Ns V_{th} \left(1 + \frac{R_s}{R_{sh}}\right)}
        \right)
 
-The the module current can be solved using the Lambert W-function.
+Then the module current can be solved using the Lambert W-function,
+:math:`W \left(z \right)`.
 
 .. math::
 
    I = \frac{I_L + I_0 - \frac{V}{R_{sh}}}{1 + \frac{R_s}{R_{sh}}}
-       - \frac{n Ns V_{th}}{R_s} W(z)
+       - \frac{n Ns V_{th}}{R_s} W \left(z \right)
 
 
 Bishop's Algorithm
 ------------------
-The function :func:`pvlib.singlediode.bishop88` uses an explicit solution [3]
+The function :func:`pvlib.singlediode.bishop88` uses an explicit solution [4]
 that finds points on the IV curve by first solving for pairs :math:`(V_d, I)`
 where :math:`V_d` is the diode voltage :math:`V_d = V + I*Rs`. Then the voltage
 is backed out from :math:`V_d`. Points with specific voltage, such as open
@@ -63,8 +67,9 @@ we can also show that when :math:`V_d = V_{oc, est}`, the resulting
 current is also negative, meaning that the corresponding voltage must be
 in the 4th quadrant and therefore greater than the open circuit voltage
 (see proof below). Therefore the entire forward-bias 1st quadrant IV-curve
-is bounded, and a bisection search within these points will always find
-desired condition.
+is bounded because :math:`V_{oc} < V_{oc, est}`, and so a bisection search
+between 0 and :math:`V_{oc, est}` will always find any desired condition in the
[…]
