### 2018-08-10 fa03a93
DOC: MAINT: fix citation for Sandia Report, remove extra parentheses
