### 2018-08-10 5e6ba7e
Fix #521: provide longitude before latitude when calling query.lonlat_box (#524)
* Fix #521: provide longitude before latitude when calling query.lonlat_box

See signature of `siphon.DataQuery.lonlat_box`: https://github.com/Unidata/siphon/blob/a6529fbaf3165f93cdffa87080295431a647df2c/siphon/http_util.py#L174

* Document fix in release notes
