### 2018-08-10 dc6bdbf
MAINT: DOC: rename singlediode module, and clean up docstring for singlediode function (#525)
* MAINT: rename singlediode_methods.py to just singlediode.py

* closes #515
* import pvlib in pvsystem and use pvlib.singlediode to differentiate
from existing pvsystem.singldiode() method already there!
* rename test_singlediode_methods.py to test_singlediode.py
* update documentation and comments

* DOC: move proof that estimated Voc is in q4

* closes #518
* add singlediode.rst and add to index

* DOC: add DOI to Jain and Kapoor

* proofed the docs and they are ok

* DOC: MAINT: improve wording in
[…]
