### 2018-08-10 d18bfce
sprinkle plt.close to reduce memory on rtd
