+++ docs/sphinx/source/clearsky.rst
@@ -280,6 +280,9 @@ Grid with a clear sky irradiance for a few turbidity values.
     @savefig ineichen-grid.png width=10in
     In [1]: plt.show();
 
+    @suppress
+    In [1]: plt.close();
+
 
 
 Validation
@@ -350,7 +353,7 @@ A clear sky time series using only basic pvlib functions.
 
     In [1]: pressure = pvlib.atmosphere.alt2pres(altitude)
 
-    In [1]: dni_extra = pvlib.irradiance.extraradiation(times)
+    In [1]: dni_extra = pvlib.irradiance.get_extra_radiation(times)
 
     # an input is a Series, so solis is a DataFrame
     In [1]: solis = clearsky.simplified_solis(apparent_elevation, aod700, precipitable_water,
@@ -366,6 +369,9 @@ A clear sky time series using only basic pvlib functions.
 
     @savefig solis-vs-time-0.1-1.png width=6in
     In [1]: plt.show();
+
+    @suppress
+    In [1]: plt.close();
 
 The input data types determine the returned output type. Array input
 results in an OrderedDict of array output, and Series input results in a
@@ -399,6 +405,9 @@ Irradiance as a function of solar elevation.
     @savefig solis-vs-elevation.png width=6in
     In [1]: ax.legend(loc=2);
 
+    @suppress
+    In [1]: plt.close();
+
 
 Grid with a clear sky irradiance for a few PW and AOD values.
 
@@ -412,7 +421,7 @@ Grid with a clear sky irradiance for a few PW and AOD values.
 
     In [1]: pressure = pvlib.atmosphere.alt2pres(altitude)
 
-    In [1]: dni_extra = pvlib.irradiance.extraradiation(times)
+    In [1]: dni_extra = pvlib.irradiance.get_extra_radiation(times)
 
     In [1]: aod700 = [0.01, 0.1]
 
@@ -428,6 +437,9 @@ Grid with a clear sky irradiance for a few PW and AOD values.
 
     @savefig solis-grid.png width=10in
     In [1]: plt.show();
+
+    @suppress
+    In [1]: plt.close();
 
 Contour plots of irradiance as a function of both PW and AOD.
 
@@ -474,15 +486,24 @@ Contour plots of irradiance as a function of both PW and AOD.
     @savefig solis-ghi.png width=10in
     In [1]: plt.show()
 
+    @suppress
+    In [1]: plt.close();
+
     In [1]: plot_solis('dni')
 
     @savefig solis-dni.png width=10in
     In [1]: plt.show()
 
+    @suppress
+    In [1]: plt.close();
+
     In [1]: plot_solis('dhi')
 
     @savefig solis-dhi.png width=10in
     In [1]: plt.show()
+
+    @suppress
+    In [1]: plt.close();
 
 
 Validation
@@ -548,6 +569,9 @@ GHI data. We first generate and plot the clear sky and measured data.
     @savefig detect-clear-ghi.png width=10in
     plt.show();
 
+    @suppress
+    plt.close();
+
 Now we run the synthetic data and clear sky estimate through the
 :py:func:`~pvlib.clearsky.detect_clearsky` function.
 
@@ -561,6 +585,9 @@ Now we run the synthetic data and clear sky estimate through the
 
     @savefig detect-clear-detected.png width=10in
     ax.set_ylabel('Clear (1) or Cloudy (0)');
+
+    @suppress
+    plt.close();
 
 The algorithm detected the cloud event and the overirradiance event.
 
