### 2018-08-10 258e32a
fix fail_on_pvlib_version
