### 2018-08-10 73c4fc5
allow xpass on python2 and windows
