+++ pvlib/test/test_conftest.py
@@ -1,9 +1,10 @@
 import pytest
 
-from conftest import fail_on_pvlib_version
+from conftest import fail_on_pvlib_version, platform_is_windows, has_python2
 
 
-@pytest.mark.xfail(strict=True,
+# allow xpass for python 2 on windows
+@pytest.mark.xfail(strict=(not (platform_is_windows and has_python2)),
                    reason='fail_on_pvlib_version should cause test to fail')
 @fail_on_pvlib_version('0.0')
 def test_fail_on_pvlib_version():
