### 2018-08-10 88e8f00
add statement about ending python 2.7 support
