### 2018-08-13 50252aa
PERF speciify time index in ModelChain.prepare_inputs
