### 2018-08-13 943cd7c
improve performance of ephemeris dataframe assembly
