+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -60,6 +60,7 @@ Enhancements
   :func:`~pvlib.singlediode.bishop88_mpp`.
 * Add PVSyst thin-film recombination losses for CdTe and a:Si (:issue:`163`)
 * Python 3.7 officially supported. (:issue:`496`)
+* Improve performance of solarposition.ephemeris. (:issue:`512`)
 
 
 Bug fixes
