### 2018-08-13 ecf2a9e
add airspeed velocity config and tests
