+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -100,6 +100,8 @@ Testing
 * Use pytest-mock to ensure that ModelChain DC model is set up correctly.
 * Add Python 3.7 to build matrix
 * Make test_forecast.py more robust. (:issue:`293`)
+* Add airspeed velocity performance testing configuration and a few tests.
+  (:issue:`419`)
 
 
 Contributors
