### 2018-08-13 #884 @sjpfenninger
It was a pleasure to review this submission. The software is well structured and the documentation is  extensive and easy to follow. Just one (minor) item not checked on the checklist: references `Holmgren2017` and `Holmgren2018` don't have a DOI - can that be remedied, or could they be linked to a PDF instead (@wholmgren)?

Some other minor comments:

* You could set the repository's website at the very top of the GitHub page to the documentation, making it quicker to get there from the repository
* Installation docs could explicitly list requirements (right now, they do so only  implicitly in the `conda create` command under "Install as an editable library", as far as I can see)
* `tracking.LocalizedSingleAxisTracker` has a somewhat mysterious docstring
* A conda-forge package would be highly appreciated :smile:
