### 2018-08-13 pvlib/pvlib-python#531 @wholmgren
improve installation docs
Installation documentation should be improved to cover:

- [x] Only pandas and numpy are required for most functionality, but additional packages can be installed to support more features. Consider documenting which features require which packages.
- [x] pvlib is available on conda-forge as ``pvlib-python``. (we should also consider changing the conda-forge package to ``pvlib`` but that is a separate issue)

This is feedback from https://github.com/openjournals/joss-reviews/issues/884#issuecomment-412553989
