### 2018-08-13 pvlib/pvlib-python#532 @wholmgren
Improve LocalizedPVSystem and LocalizedSingleAxisTracker documentation
The [LocalizedPVSystem](http://pvlib-python.readthedocs.io/en/latest/generated/pvlib.pvsystem.LocalizedPVSystem.html) and especially [LocalizedSingleAxisTracker](http://pvlib-python.readthedocs.io/en/latest/generated/pvlib.tracking.LocalizedSingleAxisTracker.html) documentation should be improved. I'd like to:

- [x] update the localized tracker docstring to look more like the localized fixed docstring
- [x] update the localized fixed and tracker docstrings to suggest that users also consider the ModelChain workflow
- [x] update the [Modeling paradigm's](http://pvlib-python.readthedocs.io/en/latest/package_overview.html#object-oriented-localizedpvsystem) text to recommend the other approaches over the localized approach because multiple inheritance is hard to do well and the localized classes may be deprecated.

Based partially on feedback in https://github.com/openjournals/joss-reviews/issues/884#issuecomment-412553989
