### 2018-08-13 27fd150
update installation instructions
