+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -87,6 +87,7 @@ Documentation
 * Expanded general contributing and pull request guidelines.
 * Added section on single diode equation with some detail on solutions used in
   pvlib-python (:issue:`518`)
+* Minor improvements and updates to installation documentation. (:issue:`531`)
 
 
 Testing
