### 2018-08-13 c086669
minor changes
