### 2018-08-14 76c1867
improve localized classes documentation
