### 2018-08-14 dafaa85
add get_[relative_airmass,sky_diffuse,ground_diffuse,extra_radiation,total_irradiance] and pvlibDeprecationWarning (#427)
* initial refactor

* remove new fields from globalinplane wrapper

* update test

* update whatsnew

* add DeprecationWarning

* update api.rst

* add _deprecation.py module

* extraradiation to get_extra_radiation

* grounddiffuse to get_ground_diffuse

* remove deprecated klutcher (misspelling) from get_sky_diffuse

* total_irrad to get_total_poa_irradiance

* update api.rst

* change airmass function names

* forgot to change get_extra_radiation in api.rst

* remove poa
[…]
