### 2018-08-14 9ead2f2
add statement about ending python 2.7 support (#527)
