### 2018-08-14 10339ee
improve performance of ephemeris dataframe assembly (#529)
