### 2018-08-14 9a74404
limit cos(aoi) in diffuse sky functions
