### 2018-08-14 c8530a2
no limit on cos_solar_zenith in hb calculation. change proj limit to -1
