### 2018-08-14 36aec41
improve localized classes documentation (#534)
