### 2018-08-14 46059b5
improve speed of Location.get_airmass (#528)
* PERF speciify time index in location.get_airmass

* PERF speciify time index in ModelChain.prepare_inputs

* update whatsnew
