### 2018-08-15 870196a
clean up whatsnew
