### 2018-08-15 18222d1
more refactoring
