### 2018-08-17 c0943a8
reimplement without projection_minimum kwarg
