+++ pvlib/irradiance.py
@@ -152,8 +152,7 @@ def _handle_extra_radiation_types(datetime_or_doy, epoch_year):
     return to_doy, to_datetimeindex, to_output
 
 
-def aoi_projection(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth,
-                   projection_minimum=-1):
+def aoi_projection(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth):
     """
     Calculates the dot product of the solar vector and the surface
     normal.
@@ -170,11 +169,6 @@ def aoi_projection(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth,
         Solar zenith angle.
     solar_azimuth : numeric
         Solar azimuth angle.
-    projection_minimum : numeric, default -1
-        Minimum allowable value for the projection for the solar angle
-        into the plane of the array. For example, 0 is equivalent to a
-        90 degree limit, 0.01745 is equivalent to a 89 degree limit, and
-        -1 allows for any angle.
 
     Returns
     -------
@@ -186,8 +180,6 @@ def aoi_projection(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth,
         tools.cosd(surface_tilt) * tools.cosd(solar_zenith) +
         tools.sind(surface_tilt) * tools.sind(solar_zenith) *
         tools.cosd(solar_azimuth - surface_azimuth))
-
-    projection = np.maximum(projection, projection_minimum)
 
     try:
         projection.name = 'aoi_projection'
@@ -234,9 +226,7 @@ def aoi(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth):
 
 
 def poa_horizontal_ratio(surface_tilt, surface_azimuth,
-                         solar_zenith, solar_azimuth,
-                         poa_projection_minimum=None,
-                         solar_zenith_projection_minimum=None):
+                         solar_zenith, solar_azimuth):
     """
     Calculates the ratio of the beam components of the plane of array
     irradiance and the horizontal irradiance.
@@ -253,12 +243,6 @@ def poa_horizontal_ratio(surface_tilt, surface_azimuth,
         Solar zenith angle.
     solar_azimuth : numeric
         Solar azimuth angle.
-    poa_projection_minimum : numeric, default -1
-        Minimum allowable value for the projection for the solar angle
-        into the plane of the array.
-    solar_zenith_projection_minimum : numeric, default -1
-        Minimum allowable value for the projection of the solar angle
-        on to a horizontal surface.
 
     Returns
     -------
@@ -268,15 +252,11 @@ def poa_horizontal_ratio(surface_tilt, surface_azimuth,
     """
 
     cos_poa_zen = aoi_projection(surface_tilt, surface_azimuth,
-                                 solar_zenith, solar_azimuth,
-                                 projection_minimum=poa_projection_minimum)
+                                 solar_zenith, solar_azimuth)
 
     cos_solar_zenith = tools.cosd(solar_zenith)
 
-    cos_solar_zenith = np.maximum(cos_solar_zenith,
-                                  solar_zenith_projection_minimum)
-
-    # ratio of titled and horizontal beam irradiance
+    # ratio of tilted and horizontal beam irradiance
     ratio = cos_poa_zen / cos_solar_zenith
 
     try:
@@ -765,8 +745,8 @@ def klucher(surface_tilt, surface_azimuth, dhi, ghi, solar_zenith,
 
     # zenith angle with respect to panel normal.
     cos_tt = aoi_projection(surface_tilt, surface_azimuth,
[…]
