### 2018-08-17 4ac71a4
refactor disc and dirint
