+++ pvlib/irradiance.py
@@ -1178,8 +1178,88 @@ def perez(surface_tilt, surface_azimuth, dhi, dni, dni_extra,
         return sky_diffuse
 
 
-def disc(ghi, zenith, datetime_or_doy, pressure=101325, min_cos_zenith=0.065,
-         kt=None):
+def clearness_index(ghi, solar_zenith, extra_radiation, min_cos_zenith=0.065):
+    """
+    Calculate the clearness index.
+
+    The clearness index is the ratio of global to extraterrestrial
+    irradiance on a horizontal plane.
+
+    Parameters
+    ----------
+    ghi : numeric
+        Global horizontal irradiance in W/m^2.
+
+    solar_zenith : numeric
+        True (not refraction-corrected) solar zenith angle in decimal
+        degrees.
+
+    extra_radiation : numeric
+        Irradiance incident at the top of the atmosphere
+
+    min_cos_zenith : numeric, default 0.065
+        Minimum value of cos(zenith) to allow when calculating global
+        clearness index `kt`. Equivalent to 86.273 degrees.
+
+    Returns
+    -------
+    kt : numeric
+        Clearness index
+
+    Reference
+    ---------
+    [1] Maxwell, E. L., "A Quasi-Physical Model for Converting Hourly
+    Global Horizontal to Direct Normal Insolation", Technical
+    Report No. SERI/TR-215-3087, Golden, CO: Solar Energy Research
+    Institute, 1987.
+    """
+    cos_zenith = tools.cosd(solar_zenith)
+    I0h = extra_radiation * np.maximum(cos_zenith, min_cos_zenith)
+    kt = ghi / I0h
+    kt = np.maximum(kt, 0)
+    kt = np.minimum(kt, 0.82)  # consider replacing with 1 or nan assignment
+    return kt
+
+
+def clearness_index_zenith_independent(clearness_index, airmass):
+    """
+    Calculate the zenith angle independent clearness index.
+
+    Parameters
+    ----------
+    clearness_index : numeric
+        Ratio of global to extraterrestrial irradiance on a horizontal
+        plane
+
+    airmass : numeric
+        Airmass
+
+    Returns
+    -------
+    kt_prime : numeric
+        Zenith independent clearness index
+
+    Reference
+    ---------
+    """
+    # Perez eqn 1
+    kt_prime = clearness_index / _kt_kt_prime_factor(airmass)
+    kt_prime = np.maximum(kt_prime, 0)
+    kt_prime = np.minimum(kt_prime, 0.82)  # From SRRL code
+    return kt_prime
+
+
+def _kt_kt_prime_factor(airmass):
+    """
+    Calculate the conversion factor between kt and kt prime.
+    Function is useful because DIRINT and GTI-DIRINT both use this.
+    """
+    # consider adding am = np.maximum(am, 12)  # GH 450
+    return 1.031 * np.exp(-1.4 / (0.9 + 9.4 / airmass)) + 0.1
+
+
+def disc(ghi, solar_zenith, datetime_or_doy, pressure=101325,
+         min_cos_zenith=0.065, max_zenith=87):
     """
     Estimate Direct Normal Irradiance from Global Horizontal Irradiance
     using the DISC model.
@@ -1203,6 +1283,14 @@ def disc(ghi, zenith, datetime_or_doy, pressure=101325, min_cos_zenith=0.065,
 
     pressure : numeric, default 101325
         Site pressure in Pascal.
+
+    min_cos_zenith : numeric, default 0.065
+        Minimum value of cos(zenith) to allow when calculating global
+        clearness index `kt`. Equivalent to 86.273 degrees.
+
+    max_zenith : numeric, default 87
[…]
