### 2018-08-17 519390a
add stubs. use zeros. more clean up
