### 2018-08-18 205cc16
fix np.minimum vs. np.maximum
