+++ pvlib/irradiance.py
@@ -746,7 +746,7 @@ def klucher(surface_tilt, surface_azimuth, dhi, ghi, solar_zenith,
     # zenith angle with respect to panel normal.
     cos_tt = aoi_projection(surface_tilt, surface_azimuth,
                             solar_zenith, solar_azimuth)
-    cos_tt = np.minimum(cos_tt, 0)  # GH 526
+    cos_tt = np.maximum(cos_tt, 0)  # GH 526
 
     F = 1 - ((dhi / ghi) ** 2)
     try:
@@ -837,7 +837,7 @@ def haydavies(surface_tilt, surface_azimuth, dhi, dni, dni_extra,
     if projection_ratio is None:
         cos_tt = aoi_projection(surface_tilt, surface_azimuth,
                                 solar_zenith, solar_azimuth)
-        cos_tt = np.minimum(cos_tt, 0)  # GH 526
+        cos_tt = np.maximum(cos_tt, 0)  # GH 526
         cos_solar_zenith = tools.cosd(solar_zenith)
         Rb = cos_tt / np.maximum(cos_solar_zenith, 0.01745)  # GH 432
     else:
@@ -934,7 +934,7 @@ def reindl(surface_tilt, surface_azimuth, dhi, dni, ghi, dni_extra,
 
     cos_tt = aoi_projection(surface_tilt, surface_azimuth,
                             solar_zenith, solar_azimuth)
-    cos_tt = np.minimum(cos_tt, 0)  # GH 526
+    cos_tt = np.maximum(cos_tt, 0)  # GH 526
 
     # do not apply cos(zen) limit here (needed for HB below)
     cos_solar_zenith = tools.cosd(solar_zenith)
