### 2018-08-18 0e3fb1a
add sky diffuse test for zenith close to 90
