### 2018-08-19 487cdb9
update installation instructions (#533)
* update installation instructions

* minor changes

* add comments

* fix line breaks
