### 2018-08-19 eead186
fix botched whatsnew merge [skip ci]
