+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -111,7 +111,6 @@ Bug fixes
   floats and NumPy arrays. (:issue:`508`)
 * Make GitHub recognize the license, add AUTHORS.md, clarify shared copyright.
   (:issue:`503`)
-* Fix argument order of longitude and latitude when querying weather forecasts by lonlat bounding box (:issue:`521`)
 * Fix issue with non-zero direct irradiance contribution to Reindl, Klucher,
   and Hay-Davies diffuse sky algorithms when the sun is behind the array.
   (:issue:`526`)
