### 2018-08-19 4f3b7e4
limit cos(aoi) in diffuse sky functions (#535)
* limit cos(aoi) in diffuse sky functions

* no limit on cos_solar_zenith in hb calculation. change proj limit to -1

* reimplement without projection_minimum kwarg

* fix np.minimum vs. np.maximum

* update tutorial

* add sky diffuse test for zenith close to 90

* fix botched whatsnew merge [skip ci]
