### 2018-08-19 #884 @ecotillasanchez
Excellent work in this contribution, a pleasure to review and test. One can easily see the comprehensive functionality and the seamless support for a variety of external sources of data. This is very valuable for benchmarking. Besides the comments raised above, I will only add one other suggestion. In the documentation, the main functionality could be structured around a tutorial or use cases, so that the user has a clear flow of what to read first and what to read next. For example, _ModelChain_ shows up nearly on top on the left panel, and then it becomes a submenu within _API reference_.
