### 2018-08-19 #884 @wholmgren
Thanks @sjpfenninger and @ecotillasanchez for your reviews!

> You could set the repository's website at the very top of the GitHub page to the documentation, making it quicker to get there from the repository

Good idea. Done.

> Installation docs could explicitly list requirements (right now, they do so only implicitly in the conda create command under "Install as an editable library", as far as I can see)

We expanded the [Compatibility](https://pvlib-python.readthedocs.io/en/latest/installation.html#compatibility) section to address this. It looks likely that we'll add a requirements.txt file to the package in the near future, so I'd like to defer a larger Installation docs reorganization until then. https://github.com/pvlib/pvlib-python/pull/533

> tracking.LocalizedSingleAxisTracker has a somewhat mysterious docstring

Fixed in https://github.com/pvlib/pvlib-python/pull/534

> A conda-forge package would be highly appreciated 😄

We've actually had one for the last ~2 years or so, but I neglected to add it to the installation documentation until now. [See here](https://pvlib-python.readthedocs.io/en/latest/installation.html#install-standard-release). https://github.com/pvlib/pvlib-python/pull/533

> the main functionality could be structured around a tutorial or use cases

@ecotillasanchez would moving [Modeling paradigms](https://pvlib-python.readthedocs.io/en/latest/package_overview.html#modeling-paradigms) out of Package Overview and into its own top-level documentation address this concern? Perhaps we could rename it something like "Intro examples" to make it more clear.

> For example, ModelChain shows up nearly on top on the left panel, and then it becomes a submenu within API reference.

In this case, the [left panel ModelChain documentation](https://pvlib-python.readthedocs.io/en/latest/modelchain.html) provides a comprehensive discussion of the feature, while the [submenu within API reference](https://pvlib-python.readthedocs.io/en/latest/api.html#modelchain) provides, well, API reference. The same situation applies to "Clear sky" and "Forecasting".

To do:

- [ ] "references Holmgren2017 and Holmgren2018 don't have a DOI - can that be remedied, or could they be linked to a PDF instead"

cc @mikofski @cwhanse
