### 2018-08-21 01039e6
copy getting support to readme
