### 2018-08-21 8737529
move modeling paradigms to intro examples
