### 2018-08-21 0be13e5
add holmgren2017 doi
