+++ paper/paper.bib
@@ -43,7 +43,8 @@
 	Author = {William F. Holmgren and Antonio T. Lorenzo and Clifford Hansen},
 	Booktitle = {2017 IEEE 44th Photovoltaic Specialists Conference},
 	Title = {A Comparison of PV Power Forecasts Using PVLib-Python},
-	Year = {2017}}
+	Year = {2017},
+	doi = {http://doi.org/10.5281/zenodo.1400857}}
 
 @inproceedings{Holmgren2018,
 	Author = {William F. Holmgren and Clifford W. Hansen and Joshua S. Stein and Mark A. Mikofski},
