### 2018-08-21 edbe4ea
copy getting support to readme (#538)
