+++ paper/paper.bib
@@ -44,13 +44,14 @@
 	Booktitle = {2017 IEEE 44th Photovoltaic Specialists Conference},
 	Title = {A Comparison of PV Power Forecasts Using PVLib-Python},
 	Year = {2017},
-	doi = {http://doi.org/10.5281/zenodo.1400857}}
+	doi = {10.5281/zenodo.1400857}}
 
 @inproceedings{Holmgren2018,
 	Author = {William F. Holmgren and Clifford W. Hansen and Joshua S. Stein and Mark A. Mikofski},
 	Booktitle = {2018 IEEE 45th Photovoltaic Specialists Conference},
 	Title = {Review of open source tools for PV modeling},
-	Year = {2018}}
+	Year = {2018},
+	doi = {10.5281/zenodo.1401378}}
 
 @article{Louwen2017,
 	Author = {Atse Louwen and Ruud E.I. Schropp and Wilfried G.J.H.M. van Sark and Andr{\'e} P.C. Faaij},
