### 2018-08-21 cac57cb
update parameters, whatsnew. [skip ci]
