+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -123,6 +123,7 @@ Bug fixes
   Hay-Davies diffuse sky algorithms. (:issue:`432`)
 * Fix argument order of longitude and latitude when querying weather forecasts
   by lonlat bounding box (:issue:`521`)
+* Fix issue with unbounded clearness index calculation in disc. (:issue:`540`)
 
 
 Documentation
