+++ pvlib/test/test_irradiance.py
@@ -2,7 +2,7 @@ import datetime
 from collections import OrderedDict
 
 import numpy as np
-from numpy import array
+from numpy import array, nan
 import pandas as pd
 
 import pytest
@@ -372,7 +372,36 @@ def test_disc_value():
 
 
 def test_disc_min_cos_zenith_max_zenith():
-    assert False
+    # map out behavior under difficult conditions with various
+    # limiting kwargs settings
+    columns = ['dni', 'kt', 'airmass']
+    times = pd.DatetimeIndex(['2016-07-19 06:11:00-0600'])
+    out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times)
+    expected = pd.DataFrame(np.array(
+        [[  0.00000000e+00,   1.16046346e-02,   3.63954476e+01]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
+
+    out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
+                          min_cos_zenith=0)
+    expected = pd.DataFrame(np.array(
+        [[  0.00000000e+00,   0.82,   3.63954476e+01]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
+
+    out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
+                          max_zenith=100)
+    expected = pd.DataFrame(np.array(
+        [[  6.68577449e+03,   1.16046346e-02,   3.63954476e+01]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
+
+    out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
+                          min_cos_zenith=0, max_zenith=100)
+    expected = pd.DataFrame(np.array(
+        [[  7.34371328e+03,   8.20000000e-01,   3.63954476e+01]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
 
 
 def test_dirint_value():
@@ -427,11 +456,29 @@ def test_dirint_coeffs():
 
 
 def test_dirint_min_cos_zenith_max_zenith():
-    assert False
+    # map out behavior under difficult conditions with various
+    # limiting kwargs settings
+    # times don't have any physical relevance
+    times = pd.DatetimeIndex(['2014-06-24T12-0700','2014-06-24T18-0700'])
+    ghi = pd.Series([0, 1], index=times)
+    solar_zenith = pd.Series([90, 89.99], index=times)
 
+    out = irradiance.dirint(ghi, solar_zenith, times)
+    expected = pd.Series([0.0, 0.0], index=times, name='dni')
+    assert_series_equal(out, expected)
 
-def test_dirindex_min_cos_zenith_max_zenith():
-    assert False
+    out = irradiance.dirint(ghi, solar_zenith, times, min_cos_zenith=0)
+    expected = pd.Series([0.0, 0.0], index=times, name='dni')
+    assert_series_equal(out, expected)
+
+    out = irradiance.dirint(ghi, solar_zenith, times, max_zenith=100)
+    expected = pd.Series([862.197596231, 848.305136439], index=times, name='dni')
+    assert_series_equal(out, expected)
+
+    out = irradiance.dirint(ghi, solar_zenith, times, min_cos_zenith=0,
+                            max_zenith=100)
+    expected = pd.Series([147655.599432, 3817.76605046], index=times, name='dni')
+    assert_series_equal(out, expected)
 
 
 def test_gti_dirint():
@@ -576,6 +623,37 @@ def test_dirindex(times):
[…]
