### 2018-08-22 5728d11
move modeling paradigms to intro examples (#539)
* move modeling paradigms to intro examples

* update whatsnew

* update text [skip ci]
