+++ docs/sphinx/source/api.rst
@@ -169,6 +169,15 @@ DNI estimation models
    irradiance.liujordan
    irradiance.gti_dirint
 
+Clearness index models
+----------------------
+
+.. autosummary::
+   :toctree: generated/
+
+   irradiance.clearness_index
+   irradiance.clearness_index_zenith_independent
+
 
 PV Modeling
 ===========
