### 2018-08-23 0abc6ce
limit pvwatts_ac to gte 0 (#542)
Closes #541
