### 2018-08-24 e58c7d5
limit clearness index in disc dirint dirindex
