+++ pvlib/test/test_irradiance.py
@@ -353,22 +353,32 @@ def test_poa_components(irrad_data, ephem_data, dni_et, relative_airmass):
     assert_frame_equal(out, expected)
 
 
-def test_disc_keys(irrad_data, ephem_data):
-    disc_data = irradiance.disc(irrad_data['ghi'], ephem_data['zenith'],
-                                ephem_data.index)
-    assert 'dni' in disc_data.columns
-    assert 'kt' in disc_data.columns
-    assert 'airmass' in disc_data.columns
-
-
 def test_disc_value():
+    columns = ['dni', 'kt', 'airmass']
     times = pd.DatetimeIndex(['2014-06-24T12-0700','2014-06-24T18-0700'])
     ghi = pd.Series([1038.62, 254.53], index=times)
     zenith = pd.Series([10.567, 72.469], index=times)
     pressure = 93193.
-    disc_data = irradiance.disc(ghi, zenith, times, pressure=pressure)
-    assert_almost_equal(disc_data['dni'].values,
-                        np.array([830.46, 676.09]), 1)
+    out = irradiance.disc(ghi, zenith, times, pressure=pressure)
+    expected = pd.DataFrame(np.array(
+        [[8.30465672e+02, 7.97424662e-01, 9.35050958e-01],
+         [6.76183401e+02, 6.37821639e-01, 3.02102114e+00]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
+
+
+def test_disc_overirradiance():
+    columns = ['dni', 'kt', 'airmass']
+    ghi = np.array([3000])
+    solar_zenith = np.full_like(ghi, 0)
+    times = pd.DatetimeIndex(start='2016-07-19 12:00:00-0600', freq='1s',
+                             periods=len(ghi))
+    out = irradiance.disc(ghi=ghi, solar_zenith=solar_zenith,
+                          datetime_or_doy=times)
+    expected = pd.DataFrame(np.array(
+        [[8.72544336e+02, 1.00000000e+00, 9.99493933e-01]]),
+        columns=columns, index=times)
+    assert_frame_equal(out, expected)
 
 
 def test_disc_min_cos_zenith_max_zenith():
@@ -378,28 +388,28 @@ def test_disc_min_cos_zenith_max_zenith():
     times = pd.DatetimeIndex(['2016-07-19 06:11:00-0600'])
     out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times)
     expected = pd.DataFrame(np.array(
-        [[  0.00000000e+00,   1.16046346e-02,   3.63954476e+01]]),
+        [[0.00000000e+00, 1.16046346e-02, 3.63954476e+01]]),
         columns=columns, index=times)
     assert_frame_equal(out, expected)
 
     out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
                           min_cos_zenith=0)
     expected = pd.DataFrame(np.array(
-        [[  0.00000000e+00,   0.82,   3.63954476e+01]]),
+        [[0.00000000e+00, 1.0, 3.63954476e+01]]),
         columns=columns, index=times)
     assert_frame_equal(out, expected)
 
     out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
                           max_zenith=100)
     expected = pd.DataFrame(np.array(
-        [[  6.68577449e+03,   1.16046346e-02,   3.63954476e+01]]),
+        [[6.68577449e+03, 1.16046346e-02, 3.63954476e+01]]),
         columns=columns, index=times)
     assert_frame_equal(out, expected)
 
     out = irradiance.disc(ghi=1.0, solar_zenith=89.99, datetime_or_doy=times,
[…]
