+++ pvlib/irradiance.py
@@ -359,7 +359,7 @@ def get_total_irradiance(surface_tilt, surface_azimuth,
         dni, ghi, dhi, dni_extra=dni_extra, airmass=airmass, model=model,
         model_perez=model_perez)
 
-    poa_ground_diffuse = get_ground_diffuse(surface_tilt, ghi, albedo,
+    poa_ground_diffuse = get_ground_diffuse(surface_tilt, ghi, albedo=albedo,
                                             surface_type)
     aoi_ = aoi(surface_tilt, surface_azimuth, solar_zenith, solar_azimuth)
     irrads = poa_components(aoi_, dni, poa_sky_diffuse, poa_ground_diffuse)
@@ -1182,7 +1182,8 @@ def perez(surface_tilt, surface_azimuth, dhi, dni, dni_extra,
         return sky_diffuse
 
 
-def clearness_index(ghi, solar_zenith, extra_radiation, min_cos_zenith=0.065):
+def clearness_index(ghi, solar_zenith, extra_radiation, min_cos_zenith=0.065,
+                    max_clearness_index=2.0):
     """
     Calculate the clearness index.
 
@@ -1204,6 +1205,11 @@ def clearness_index(ghi, solar_zenith, extra_radiation, min_cos_zenith=0.065):
     min_cos_zenith : numeric, default 0.065
         Minimum value of cos(zenith) to allow when calculating global
         clearness index `kt`. Equivalent to zenith = 86.273 degrees.
+
+    max_clearness_index : numeric, default 2.0
+        Maximum value of the clearness index. The default, 2.0, allows
+        for over-irradiance events typically seen in sub-hourly data.
+        NREL's SRRL Fortran code used 0.82 for hourly data.
 
     Returns
     -------
@@ -1227,11 +1233,12 @@ def clearness_index(ghi, solar_zenith, extra_radiation, min_cos_zenith=0.065):
     kt = np.maximum(kt, 0)
     # Limit copied from the kt prime limit in dirint, which was justified
     # with reference to SRRL code. consider replacing with 1 or nan
-    kt = np.minimum(kt, 0.82)
+    kt = np.minimum(kt, max_clearness_index)
     return kt
 
 
-def clearness_index_zenith_independent(clearness_index, airmass):
+def clearness_index_zenith_independent(clearness_index, airmass,
+                                       max_clearness_index=2.0):
     """
     Calculate the zenith angle independent clearness index.
 
@@ -1243,6 +1250,11 @@ def clearness_index_zenith_independent(clearness_index, airmass):
 
     airmass : numeric
         Airmass
+
+    max_clearness_index : numeric, default 2.0
+        Maximum value of the clearness index. The default, 2.0, allows
+        for over-irradiance events typically seen in sub-hourly data.
+        NREL's SRRL Fortran code used 0.82 for hourly data.
 
     Returns
     -------
@@ -1258,7 +1270,7 @@ def clearness_index_zenith_independent(clearness_index, airmass):
     # Perez eqn 1
     kt_prime = clearness_index / _kt_kt_prime_factor(airmass)
     kt_prime = np.maximum(kt_prime, 0)
-    kt_prime = np.minimum(kt_prime, 0.82)  # From SRRL code
+    kt_prime = np.minimum(kt_prime, max_clearness_index)
     return kt_prime
 
 
@@ -1281,6 +1293,8 @@ def disc(ghi, solar_zenith, datetime_or_doy, pressure=101325,
     The DISC algorithm converts global horizontal irradiance to direct
     normal irradiance through empirical relationships between the global
     and direct clearness indices.
+
[…]
