+++ paper/paper.bib
@@ -75,7 +75,8 @@
 	Author = {Mark A. Mikofski and Clifford W. Hansen and William F. Holmgren and Gregory M. Kimball},
 	Booktitle = {2017 IEEE 44th Photovoltaic Specialists Conference},
 	Title = {Use of Measured Aerosol Optical Depth and Precipitable Water to Model Clear Sky Irradiance},
-	Year = {2017}}
+	Year = {2017},
+	doi = {10.5281/zenodo.1403238}}
 
 @article{Polo2016,
 	Author = {J. Polo, S. Garcia-Bouhaben, M. C. Alonso-Garcia},
