### 2018-08-25 f825805
update pvlib doi [skip ci]
