### 2018-08-25 845a75a
update joss paper DOIs (#544)
* add joss badge [skip ci]

* add holmgren2017 doi

* add doi [skip ci]

* add mikofski2017

* update pvlib doi [skip ci]
