### 2018-08-25 #884 @wholmgren
@sjpfenninger @ecotillasanchez I believe we've addressed all of your helpful comments. Let us know if we're missing anything. Thanks!
