### 2018-08-25 f6a35a1
improve atmo tests. fix youngirvine1967
