### 2018-08-25 152bc64
change pandas testing import
