### 2018-08-25 b3f43cb
maybe fix failing py27 test
