### 2018-08-25 9eeb67f
doc clean up
