### 2018-08-27 23b97c2
Add GTI DIRINT model (#400)
* implement gtidirint without iteration

* add option to skip aoi > 90

* make gti dirint gte 90 function private

* add some gti_dirint tests

* implement vectorized iterative solution. fix gt 90 error

* use azimuth in gt 90 function

* fix some issues near 90 deg

* remove unnecessary comment

* add kt option to disc, dirint to fix gte 90 dni issue

* update whatsnew, add to api.rst

* move to 0.5.2 whatsnew

* add gti_dirint tutorial

* move changes to 0.6.0 whatsnew

* update parameters

* fix 0.5.2 whatsnew

* rework max_iterations logic

* flake8

* update tu
[…]
