### 2018-08-27 e3faaad
fix bug in get_relative_airmass(model=youngirvine1967). improve test_atmosphere (#546)
* ignore vscode

* improve atmo tests. fix youngirvine1967

* change pandas testing import
