### 2018-08-27 #884 @katyhuff
Thanks for quickly handling all of the reviewer comments @wholmgren ! It looks to me like the comments from @sjpfenninger and @ecotillasanchez have been handled. @sjpfenninger and @ecotillasanchez can you confirm that you approve these changes? Once you confrim, we'll be ready to move forward with acceptance.
