### 2018-08-27 d3c1ae3
add consistency tests to ModelChain.dc_model
