### 2018-08-27 6884555
Add self argument
