+++ pvlib/modelchain.py
@@ -402,17 +402,17 @@ class ModelChain(object):
         else:
             self._dc_model = partial(model, self)
 
-    def _sapm_param_set():
+    def _sapm_param_set(self):
         return set(['A0', 'A1', 'A2', 'A3', 'A4', 'B0', 'B1', 'B2', 'B3', 'B4', 
                     'B5', 'C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'Isco',
                     'Impo', 'Aisc', 'Aimp', 'Bvoco', 'Mbvoc', 'Bvmpo', 'Mbvmp',
                     'N', 'Cells_in_Series', 'IX0', 'IXX0', 'FD'])
 
-    def _singlediode_param_set():
+    def _singlediode_param_set(self):
         return set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref', 'R_sh_ref',
                     'R_s'])
             
-    def _pvwatts_dc_param_set():
+    def _pvwatts_dc_param_set(self):
         return set(['pdc0', 'gamma_pdc'])
 
     def infer_dc_model(self):
