### 2018-08-28 1c4ac04
Move DC_MODEL_PARAMS to pvsystem, streamline validation
