### 2018-08-28 a04e111
Fix key for DC_MODEL_PARAMS
