+++ pvlib/pvsystem.py
@@ -30,7 +30,7 @@ DC_MODEL_PARAMS = {'sapm' :
                           'B4', 'B5', 'C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6',
                           'C7', 'Isco', 'Impo', 'Aisc', 'Aimp', 'Bvoco',
                           'Mbvoc', 'Bvmpo', 'Mbvmp', 'N', 'Cells_in_Series',
-                          'IX0', 'IXX0', 'FD']),
+                          'IXO', 'IXXO', 'FD']),
                    'singlediode' :
                      set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
                           'R_sh_ref', 'R_s']),
