### 2018-08-28 1fa1be3
Change logic in dc_model, correct pvwatts keyword
