+++ pvlib/modelchain.py
@@ -361,17 +361,20 @@ class ModelChain(object):
 
     @dc_model.setter
     def dc_model(self, model):
+        # guess at model if None
         if model is None:
-            self._dc_model = self.infer_dc_model()
-        elif isinstance(model, str):
+            self._dc_model, model = self.infer_dc_model()
+
+        # Set model and validate parameters
+        if isinstance(model, str):
             model = model.lower()
             # validate module parameters
             missing_params = DC_MODEL_PARAMS[model] - \
                                     set(self.system.module_parameters.keys())
             if missing_params: # some parameters are not in module.keys()
                 raise ValueError(model + ' selected for the DC model but '
-                                     'the module is missing one or more '
-                                     'required parameters: ' +
+                                     'one or more required parameters '
+                                     ' are missing : ' +
                                      str(missing_params))
             if model == 'sapm':
                 self._dc_model = self.sapm
@@ -385,13 +388,16 @@ class ModelChain(object):
             self._dc_model = partial(model, self)
 
     def infer_dc_model(self):
+        # returns both model function object and model string, could drop
+        # model function object in the future since the model function object
+        # will be set in dc_model after validating parameter consistency
         params = set(self.system.module_parameters.keys())
         if set(['A0', 'A1', 'C7']) <= params:
-            return self.sapm
+            return self.sapm, 'sapm'
         elif set(['a_ref', 'I_L_ref', 'I_o_ref', 'R_sh_ref', 'R_s']) <= params:
-            return self.singlediode
+            return self.singlediode, 'singlediode'
         elif set(['pdc0', 'gamma_pdc']) <= params:
-            return self.pvwatts_dc
+            return self.pvwatts_dc, 'pvwatts_dc'
         else:
             raise ValueError('could not infer DC model from '
                              'system.module_parameters')
