### 2018-08-28 971f6d7
use rst instead of ipynb
