### 2018-08-28 121a8b1
More adjustment to pvwatts_dc vs pvwatts keyword
