+++ docs/sphinx/source/pvsystem.rst
@@ -1,12 +1,20 @@
-.. _pvsystem:
+.. _pvsystemdoc:
 
 PVSystem
 ========
 
-The :py:class:`~.pvsystem.PVSystem` class wraps many of the functions in
-the :py:mod:`~.pvsystem` module. This simplifies the API by eliminating
-the need for a user to specify arguments such as module and
-inverter properties when calling PVSystem methods.
+.. ipython:: python
+    :suppress:
+
+    from pvlib import pvsystem
+
+The :py:class:`~pvlib.pvsystem.PVSystem` class wraps many of the
+functions in the :py:mod:`~pvlib.pvsystem` module. This simplifies the
+API by eliminating the need for a user to specify arguments such as
+module and inverter properties when calling PVSystem methods.
+:py:class:`~pvlib.pvsystem.PVSystem` is not better or worse than the
+functions it wraps -- it is simply an alternative way of organizing
+your data.
 
 This guide aims to build understanding of the PVSystem class. It assumes
 basic familiarity with object-oriented code in Python, but most
@@ -14,8 +22,8 @@ information should be understandable without a solid understanding of
 classes. Keep in mind that `functions` are independent of objects,
 while `methods` are attached to objects.
 
-See the :ref:`modelchain` documentation for a similar guide
-focused on using PVSystem objects in time series modeling applications.
+See :py:class:`~pvlib.modelchain.ModelChain` for an application of
+PVSystem to time series modeling.
 
 
 .. _designphilosophy:
@@ -34,39 +42,41 @@ Intrinsic data is stored in object attributes. For example, the data
 that describes a PV system's module parameters is stored in
 `PVSystem.module_parameters`.
 
-.. ipython::
+.. ipython:: python
 
     module_parameters = {'pdc0': 10, 'gamma_pdc': -0.004}
     system = pvsystem.PVSystem(module_parameters=module_parameters)
     print(system.module_parameters)
 
 Extrinsic data is passed to a PVSystem as method arguments. For example,
-the :py:meth:`~pvsystem.PVSystem.pvwatts_dc` method accepts extrinsic
+the :py:meth:`~pvlib.pvsystem.PVSystem.pvwatts_dc` method accepts extrinsic
 data irradiance and temperature.
 
-.. ipython::
+.. ipython:: python
 
     pdc = system.pvwatts_dc(1000, 30)
     print(pdc)
 
-Compare the :py:meth:`~pvsystem.PVSystem.pvwatts_dc` method signature
-to the :py:func:`~pvsystem.pvwatts_dc` function signature.
+Compare the :py:meth:`~pvlib.pvsystem.PVSystem.pvwatts_dc` method signature
+to the :py:func:`~pvlib.pvsystem.pvwatts_dc` function signature:
 
-:py:meth:`~pvsystem.PVSystem.pvwatts_dc`: ``pvwatts_dc(g_poa_effective, temp_cell)``
-:py:func:`~pvsystem.pvwatts_dc`: ``pvwatts_dc(g_poa_effective, temp_cell, pdc0, gamma_pdc, temp_ref=25.)``
+    * :py:meth:`PVSystem.pvwatts_dc(g_poa_effective, temp_cell) <pvlib.pvsystem.PVSystem.pvwatts_dc>`
+    * :py:func:`pvwatts_dc(g_poa_effective, temp_cell, pdc0, gamma_pdc, temp_ref=25.) <pvlib.pvsystem.pvwatts_dc>`
 
-How does this work? The :py:meth:`~pvsystem.PVSystem.pvwatts_dc` method
-looks in `PVSystem.module_parameters` for the `pdc0`, `gamma_pdc`
-arguments. Then the :py:meth:`~pvsystem.PVSystem.pvwatts_dc` calls the
[…]
