### 2018-08-29 10a5f7d
Add handling of invalid model name string
