### 2018-08-30 c0f58c3
Add test for invalid dc model parameters
