### 2018-08-30 e11d1c5
Add to whatsnew.rst
