### 2018-08-30 743c621
Cleanup, remove copy from test
