+++ pvlib/modelchain.py
@@ -375,7 +375,7 @@ class ModelChain(object):
                 if missing_params: # some parameters are not in module.keys()
                     raise ValueError(model + ' selected for the DC model but '
                                          'one or more required parameters '
-                                         ' are missing : ' +
+                                         'are missing : ' +
                                          str(missing_params))
                 if model == 'sapm':
                     self._dc_model = self.sapm
