### 2018-08-30 669d4bd
Delete comment from infer_dc_model
