### 2018-08-30 0036df5
add scale section. other edits
