### 2018-08-31 abe8bf1
add consistency tests to ModelChain.dc_model (#548)
* add consistency tests to ModelChain.dc_model

* Add self argument

* Move DC_MODEL_PARAMS to pvsystem, streamline validation

* Fix key for DC_MODEL_PARAMS

* Fix typo in IXO, IXXO

* Change logic in dc_model, correct pvwatts keyword

* More adjustment to pvwatts_dc vs pvwatts keyword

* Add handling of invalid model name string

* Add test for invalid dc model parameters

* Fix typo in cec_dc_adr_ac_system

* Add to whatsnew.rst

* Cleanup, remove copy from test

* Delete comment from infer_dc_model
