+++ docs/sphinx/source/index.rst
@@ -75,6 +75,7 @@ Contents
    whatsnew
    installation
    contributing
+   pvsystem
    modelchain
    timetimezones
    clearsky
