### 2018-09-04 b4737d3
Add desoto and pvsyst as dc_model options to ModelChain
