+++ pvlib/pvsystem.py
@@ -34,6 +34,10 @@ DC_MODEL_PARAMS = {'sapm' :
                    'singlediode' :
                      set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
                           'R_sh_ref', 'R_s']),
+                   'pvsyst' :
+                     set(['gamma_ref', 'mu_gamma', 'I_L_ref', 'I_o_ref',
+                          'R_sh_ref', 'R_sh_0', 'R_s', 'alpha_sc', 'EgRef',
+                          'cells_in_series']),
                    'pvwatts' :
                      set(['pdc0', 'gamma_pdc'])
                   }
