### 2018-09-04 65938f7
BUG: fix brentq doesn't raise error in singlediode
* if `brentq` is called in `singlediode.py` but no SciPy, then should raise an exception
* currently using `NotImplemented` raises TypeError: 'NotImplementedType' object is not callable because `NotImplemented` is a constant
