### 2018-09-04 a5476e5
fix variable/column name bug in HRRR_ESRL.output_variables
