### 2018-09-04 e7fb500
Add desoto to DC_MODEL_PARAMS
