### 2018-09-04 4c621c6
Change to pvlibDeprecationWarning, add v0.7 test for singlediode keyword
