### 2018-09-04 9777ba6
Fix model function reference in test_infer_dc_model
