### 2018-09-04 f2d8fe1
Change function object to strings
