### 2018-09-04 3a3afc6
Add pvsyst_dc_snl_ac_system to test_infer_dc_models argument list
