### 2018-09-04 df3a55f
Add sam_data argument to pvsyst_dc_snl_ac_system
