### 2018-09-04 7166fb0
fix variable/column name bug in HRRR_ESRL.output_variables (#558)
