### 2018-09-05 219e8f4
Update whatsnew, api.rst
