### 2018-09-05 7bf8a30
Add deprecated to ModelChain.singlediode docstring
