### 2018-09-05 b5ed45f
Revert "Add deprecated to ModelChain.singlediode docstring"
This reverts commit 7bf8a30af0b8e2c526ab894a11d7127bb9f151c4.
