### 2018-09-05 85e9aac
Add deprecated to ModelChain.singlediode docstring
