### 2018-09-05 230349c
Fix indents, remove ModelChain.singlediode from api.rst
