+++ pvlib/singlediode.py
@@ -155,7 +155,7 @@ def bishop88(diode_voltage, photocurrent, saturation_current,
         grad_i_recomb = np.where(is_recomb, i_recomb / v_recomb, 0)
         grad_2i_recomb = np.where(is_recomb, 2 * grad_i_recomb / v_recomb, 0)
         g_diode = saturation_current * np.exp(v_star) / nNsVth  # conductance
-        grad_i = -g_diode - g_sh - grad_i_recomb # di/dvd
+        grad_i = -g_diode - g_sh - grad_i_recomb  # di/dvd
         grad_v = 1.0 - grad_i * resistance_series  # dv/dvd
         # dp/dv = d(iv)/dv = v * di/dv + i
         grad = grad_i / grad_v  # di/dv
@@ -468,7 +468,7 @@ def _lambertw_v_from_i(resistance_shunt, resistance_series, nNsVth, current,
         #  V = -I*(Rs + Rsh) + IL*Rsh - a*lambertwterm + I0*Rsh
         # Recast in terms of Gsh=1/Rsh for better numerical stability.
         V[idx_p] = (IL[idx_p] + I0[idx_p] - I[idx_p]) / Gsh[idx_p] - \
-                   I[idx_p] * Rs[idx_p] - a[idx_p] * lambertwterm
+            I[idx_p] * Rs[idx_p] - a[idx_p] * lambertwterm
 
     if output_is_scalar:
         return np.asscalar(V)
@@ -500,7 +500,7 @@ def _lambertw_i_from_v(resistance_shunt, resistance_series, nNsVth, voltage,
                             voltage, saturation_current, photocurrent)
 
     # Intitalize output I (V might not be float64)
-    I = np.full_like(V, np.nan, dtype=np.float64)
+    I = np.full_like(V, np.nan, dtype=np.float64)           # noqa: E741, N806
 
     # Determine indices where 0 < Rs requires implicit model solution
     idx_p = 0. < Rs
@@ -540,7 +540,7 @@ def _lambertw_i_from_v(resistance_shunt, resistance_series, nNsVth, voltage,
 
 
 def _lambertw(photocurrent, saturation_current, resistance_series,
-                         resistance_shunt, nNsVth, ivcurve_pnts=None):
+              resistance_shunt, nNsVth, ivcurve_pnts=None):
     # Compute short circuit current
     i_sc = _lambertw_i_from_v(resistance_shunt, resistance_series, nNsVth, 0.,
                               saturation_current, photocurrent)
@@ -593,7 +593,7 @@ def _pwr_optfcn(df, loc):
     Function to find power from ``i_from_v``.
     '''
 
-    I = _lambertw_i_from_v(df['r_sh'], df['r_s'], df['nNsVth'], df[loc],
-                           df['i_0'], df['i_l'])
+    I = _lambertw_i_from_v(df['r_sh'], df['r_s'],           # noqa: E741, N806
+                           df['nNsVth'], df[loc], df['i_0'], df['i_l'])
 
     return I * df[loc]
