### 2018-09-05 cba621b
Add desoto and pvsyst as dc_model options to ModelChain (#555)
* Add desoto and pvsyst as dc_model options to ModelChain

* Add desoto to DC_MODEL_PARAMS

* Change to pvlibDeprecationWarning, add v0.7 test for singlediode keyword

* Fix model function reference in test_infer_dc_model

* Change function object to strings

* Add pvsyst_dc_snl_ac_system to test_infer_dc_models argument list

* Add sam_data argument to pvsyst_dc_snl_ac_system

* Update whatsnew, api.rst

* Add deprecated to ModelChain.singlediode docstring

* Revert "Add deprecated to ModelChain.singlediode docstring"

This revert
[…]
