### 2018-09-05 20ce777
Implement CEC model
