+++ pvlib/pvsystem.py
@@ -34,6 +34,9 @@ DC_MODEL_PARAMS = {'sapm' :
                    'desoto' :
                      set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
                           'R_sh_ref', 'R_s']),
+                   'cec' :
+                     set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
+                          'R_sh_ref', 'R_s', 'Adjust']),
                    'pvsyst' :
                      set(['gamma_ref', 'mu_gamma', 'I_L_ref', 'I_o_ref',
                           'R_sh_ref', 'R_sh_0', 'R_s', 'alpha_sc', 'EgRef',
@@ -338,6 +341,35 @@ class PVSystem(object):
                                 self.module_parameters)
 
         return calcparams_desoto(effective_irradiance, temp_cell, **kwargs)
+
+    def calcparams_cec(self, effective_irradiance, temp_cell, **kwargs):
+        """
+        Use the :py:func:`calcparams_cec` function, the input
+        parameters and ``self.module_parameters`` to calculate the
+        module currents and resistances.
+
+        Parameters
+        ----------
+        effective_irradiance : numeric
+            The irradiance (W/m2) that is converted to photocurrent.
+
+        temp_cell : float or Series
+            The average cell temperature of cells within a module in C.
+
+        **kwargs
+            See pvsystem.calcparams_cec for details
+
+        Returns
+        -------
+        See pvsystem.calcparams_cec for details
+        """
+
+        kwargs = _build_kwargs(['a_ref', 'I_L_ref', 'I_o_ref', 'R_sh_ref',
+                                'R_s', 'alpha_sc', 'Adjust', 'EgRef', 'dEgdT',
+                                'irrad_ref', 'temp_ref'],
+                                self.module_parameters)
+
+        return calcparams_cec(effective_irradiance, temp_cell, **kwargs)
 
     def calcparams_pvsyst(self, effective_irradiance, temp_cell):
         """
@@ -1225,6 +1257,146 @@ def calcparams_desoto(effective_irradiance, temp_cell,
     Rs = R_s
 
     return IL, I0, Rs, Rsh, nNsVth
+
+
+def calcparams_cec(effective_irradiance, temp_cell,
+                   alpha_sc, a_ref, I_L_ref, I_o_ref, R_sh_ref, R_s,
+                   Adjust, EgRef=1.121, dEgdT=-0.0002677,
+                   irrad_ref=1000, temp_ref=25):
+    '''
+    Calculates five parameter values for the single diode equation at
+    effective irradiance and cell temperature using the CEC
+    model described in [1]. The five values returned by calcparams_cec
+    can be used by singlediode to calculate an IV curve.
+
+    Parameters
+    ----------
+    effective_irradiance : numeric
+        The irradiance (W/m2) that is converted to photocurrent.
+
+    temp_cell : numeric
+        The average cell temperature of cells within a module in C.
+
+    alpha_sc : float
+        The short-circuit current temperature coefficient of the
+        module in units of A/C.
+
+    a_ref : float
+        The product of the usual diode ideality factor (n, unitless),
+        number of cells in series (Ns), and cell thermal voltage at reference
+        conditions, in units of V.
+
+    I_L_ref : float
+        The light-generated current (or photocurrent) at reference conditions,
+        in amperes.
+
+    I_o_ref : float
+        The dark or diode reverse saturation current at reference conditions,
+        in amperes.
+
+    R_sh_ref : float
+        The shunt resistance at reference conditions, in ohms.
+
+    R_s : float
+        The series resistance at reference conditions, in ohms.
+
+    Adjust : float
+        The adjustment to the temperature coefficient for short circuit 
+        current, in percent
+
+    EgRef : float
[…]
