### 2018-09-05 bc0ae7d
STY: BUG: fix pvsystem imports pvlib twice
* from pvlib import singlediode as _singlediode # to differentiate from
local function pvsystem.singlediode
* replace pvlib.singlediode with _singlediode everywhere
* remove extra line and whitespace around FC_MODEL_PARAMS
