+++ pvlib/pvsystem.py
@@ -19,31 +19,29 @@ import pandas as pd
 from pvlib import tools
 from pvlib.tools import _build_kwargs
 from pvlib.location import Location
-from pvlib import irradiance, atmosphere
-import pvlib  # use pvlib.singlediode to avoid clash with local method
+from pvlib import irradiance, atmosphere, singlediode as _singlediode
 
 
 # a dict of required parameter names for each DC power model
-
-DC_MODEL_PARAMS = {'sapm' :
-                     set(['A0', 'A1', 'A2', 'A3', 'A4', 'B0', 'B1', 'B2', 'B3',
-                          'B4', 'B5', 'C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6',
-                          'C7', 'Isco', 'Impo', 'Aisc', 'Aimp', 'Bvoco',
-                          'Mbvoc', 'Bvmpo', 'Mbvmp', 'N', 'Cells_in_Series',
-                          'IXO', 'IXXO', 'FD']),
-                   'desoto' :
-                     set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
-                          'R_sh_ref', 'R_s']),
-                   'pvsyst' :
-                     set(['gamma_ref', 'mu_gamma', 'I_L_ref', 'I_o_ref',
-                          'R_sh_ref', 'R_sh_0', 'R_s', 'alpha_sc', 'EgRef',
-                          'cells_in_series']),
-                   'singlediode' :
-                     set(['alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
-                          'R_sh_ref', 'R_s']),
-                   'pvwatts' :
-                     set(['pdc0', 'gamma_pdc'])
-                  }
+DC_MODEL_PARAMS = {
+    'sapm' : set([
+        'A0', 'A1', 'A2', 'A3', 'A4', 'B0', 'B1', 'B2', 'B3',
+        'B4', 'B5', 'C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6',
+        'C7', 'Isco', 'Impo', 'Aisc', 'Aimp', 'Bvoco',
+        'Mbvoc', 'Bvmpo', 'Mbvmp', 'N', 'Cells_in_Series',
+        'IXO', 'IXXO', 'FD']),
+    'desoto' : set([
+        'alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
+        'R_sh_ref', 'R_s']),
+    'pvsyst' : set([
+        'gamma_ref', 'mu_gamma', 'I_L_ref', 'I_o_ref',
+        'R_sh_ref', 'R_sh_0', 'R_s', 'alpha_sc', 'EgRef',
+        'cells_in_series']),
+    'singlediode' : set([
+        'alpha_sc', 'a_ref', 'I_L_ref', 'I_o_ref',
+        'R_sh_ref', 'R_s']),
+    'pvwatts' : set(['pdc0', 'gamma_pdc'])
+}
 
 
 # not sure if this belongs in the pvsystem module.
@@ -2019,7 +2017,7 @@ def singlediode(photocurrent, saturation_current, resistance_series,
     # Calculate points on the IV curve using the LambertW solution to the
     # single diode equation
     if method.lower() == 'lambertw':
-        out = pvlib.singlediode._lambertw(
+        out = _singlediode._lambertw(
             photocurrent, saturation_current, resistance_series,
             resistance_shunt, nNsVth, ivcurve_pnts
         )
@@ -2032,19 +2030,19 @@ def singlediode(photocurrent, saturation_current, resistance_series,
         # equation for the diode voltage V_d then backing out voltage
         args = (photocurrent, saturation_current, resistance_series,
[…]
