### 2018-09-05 05b0674
follow up LGTM integration (#561)
* add .lgtm.yml file

* update whatsnew
