+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -166,6 +166,7 @@ Testing
 * Add Python 3.7 to build matrix
 * Make test_forecast.py more robust. (:issue:`293`)
 * Improve test_atmosphere.py. (:issue:`158`)
+* Add LGTM.com integration. (:issue:`554`)
 
 
 Contributors
@@ -175,4 +176,4 @@ Contributors
 * Cliff Hansen
 * Mark Mikofski
 * Alan Mathew
-
+* Xavier Rene-Corail
