### 2018-09-06 3ba597e
DOC: add example for Kasten pyrheliometric formula
* closes #302
* add references to Kasten 96 and Molineaux
* discuss using Kasten and calculating broadband AOD usinge Angstrom turbidity model

Signed-off-by: Mark Mikofski <bwana.marko@yahoo.com>
