### 2018-09-06 fffab3a
DOC: update what's new, add refs to Bird & Hulstrom and Angstrom
