+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -167,6 +167,7 @@ Testing
 * Make test_forecast.py more robust. (:issue:`293`)
 * Improve test_atmosphere.py. (:issue:`158`)
 * Add LGTM.com integration. (:issue:`554`)
+* Add SticklerCI integration.
 
 
 Contributors
