### 2018-09-06 1f62fe5
Add cec to test_infer_dc_model
