### 2018-09-06 f16e53e
style fixes (#559)
* style fixes

* more style fixes

* clean up merge
