### 2018-09-06 e149e3e
DOC: split sentences, replace abbrev. AOD
* use TMY in pvlib/data, lowercase MBARS, use non-specific variable names
* split output from readtmy3 instead of using indexing for clarity
