### 2018-09-06 #884 @katyhuff
Ah -- I'm so sorry wholmgren  -- I missed the last notification.

Indeed, it's time for the next step!

Could you please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission! (You may want to be sure to double check the paper, all minor details, etc. before creating the archive).
