### 2018-09-06 e2c28ca
change irradiance.perez(return_components=True) to return dict/DataFrame
