### 2018-09-06 d4455a1
Adjust test_infer_dc_model for desoto, singlediode cases
