+++ pvlib/test/test_modelchain.py
@@ -240,6 +240,9 @@ def test_infer_dc_model(system, cec_dc_snl_ac_system, pvsyst_dc_snl_ac_system,
                          'singlediode': 'calcparams_desoto',
                          'pvwatts_dc': 'pvwatts_dc'}
     system = dc_systems[dc_model]
+    # remove Adjust from model parameters for desoto, singlediode
+    if dc_model in ['desoto', 'singlediode']:
+        system.module_parameters.pop('Adjust')
     m = mocker.spy(system, dc_model_function[dc_model])
     mc = ModelChain(system, location,
                     aoi_model='no_loss', spectral_model='no_loss')
