### 2018-09-06 4b93e22
DOC: fix up ipython example
