### 2018-09-06 8f3004e
DOC: split sentences, revise, add angstrom example
