### 2018-09-06 0a18b42
Reapply edits to DC_MODEL_PARAMS
