### 2018-09-06 978ca5c
assign series name in test
