### 2018-09-06 #884 @wholmgren
Thanks. Looks like we have some issues with reference formatting. I'll try to fix those. Is it possible to run the pdf compilation locally without too much trouble?

We are very close to a new release so maybe best to wait a few days for that DOI.
