### 2018-09-06 daf04c5
ENH: set default alpha to 1.14, update what's new
