+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -107,6 +107,7 @@ Enhancements
 * Add irradiance.clearness_index_zenith_independent function. (:issue:`396`)
 * Add checking for consistency between module_parameters and dc_model. (:issue:`417`)
 * Add DC model methods desoto and pvsyst to ModelChain (:issue:`487`)
+* Set default alpha to 1.14 in :func:`~pvlib.atmosphere.angstrom_aod_at_lambda` (:issue:`563`)
 
 
 Bug fixes
