### 2018-09-06 609bb60
fix py27min test attempt 2
