### 2018-09-06 f64f718
DOC: use coerce year, use v0.6.0 get_property() for airmass
