### 2018-09-06 70231ea
DOC: print calculated AOD, move TL calculation after turbidity section
