### 2018-09-06 04fd379
DOC: wrap long lines, abbrev AOD, be explicit re: fictitious AOD
