### 2018-09-07 f427a39
add extras_require complete and test options
