+++ docs/sphinx/source/whatsnew/v0.6.0.rst
@@ -107,6 +107,9 @@ Enhancements
 * Add irradiance.clearness_index_zenith_independent function. (:issue:`396`)
 * Add checking for consistency between module_parameters and dc_model. (:issue:`417`)
 * Add DC model methods desoto and pvsyst to ModelChain (:issue:`487`)
+* pvlib now ships with a pvlib[complete] installation option to automatically
+  install all packages needed to support all pvlib features:
+  ``pip install pvlib[complete]`` (:issue:`553`, :issue:`483`)
 
 
 Bug fixes
