### 2018-09-07 #884 @katyhuff
Hi @wholmgren the references look ok to me. Perhaps I'm not sufficiently discerning. Can you give an example of what the problem is in the reference formatting?

Regarding the DOI, technically we'd like the DOI to be representative of the reviewed version of the code. However, if it's really just a few days to release, I think waiting for the updated doi is fine as long as new changes in your upcoming release don't alter the functionality we reviewed, do contain their own appropriate test coverage, and don't challenge any of the other elements necessary for successful JOSS review (docs, etc). I think with the development workflow you have demonstrated, this shouldn't be a problem for this project.
