### 2018-09-07 #884 @wholmgren
Thanks @katyhuff. I believe that the reviewers ultimately approved commit [845a75](https://github.com/pvlib/pvlib-python/tree/845a75ac736c00901c5077f2e57902d505e18d71). I could describe this as v0.6.0-alpha and follow the normal GitHub Release --> Zenodo archive process. Alternatively, I could download the repo at that commit and upload it as a separate Zenodo archive. Is one approach preferable to the other?

Re references, I was expecting all citations in the text to be rendered as e.g. Lastname (Year), Lastname and Lastname (Year), or Lastname et al (Year). I was confused about the presence of initials and especially full first names. Maybe it's because many of our citations are proceedings. In any case, it's fine with me if it's fine with you. The References section looks ok.
