### 2018-09-07 ec329e2
DOC: Discuss Angstrom Kasten TL, set default alpha (#563)
* DOC: add example for Kasten pyrheliometric formula

* closes #302
* add references to Kasten 96 and Molineaux
* discuss using Kasten and calculating broadband AOD usinge Angstrom turbidity model

Signed-off-by: Mark Mikofski <bwana.marko@yahoo.com>

* DOC: update what's new, add refs to Bird & Hulstrom and Angstrom

* DOC: split sentences, replace abbrev. AOD

* use TMY in pvlib/data, lowercase MBARS, use non-specific variable names
* split output from readtmy3 instead of using indexing for clarity

* DOC: fix up ipython example

* DO
[…]
