### 2018-09-07 #884 @katyhuff
- Both DOI generation approaches are equivalent, from the perspective of JOSS.
- Regarding the references -- I see. Indeed, I think that's happening at the level of the bibstyle, so it's right according to JOSS.

Thanks!
