### 2018-09-07 ff24cc8
Add test for pvsystem.calcparams_cec, extend test for calcparams_desoto, update whatnews.rst and api.rst
