+++ pvlib/pvsystem.py
@@ -1262,8 +1262,8 @@ def calcparams_cec(effective_irradiance, temp_cell,
     '''
     Calculates five parameter values for the single diode equation at
     effective irradiance and cell temperature using the CEC
-    model described in [1]. The CEC model differs from the De soto et al. 
-    model [3] by the parameter Adjust. The five values returned by 
+    model described in [1]. The CEC model differs from the De soto et al.
+    model [3] by the parameter Adjust. The five values returned by
     calcparams_cec can be used by singlediode to calculate an IV curve.
 
     Parameters
