### 2018-09-07 f303666
Replace np.round with check_less_precise in pd.Series tests
